core, special_install: more robust way of getting the correct osg-release RPM, and some other fixups
Take advantage of Python 3 features: FileNotFoundError, subprocess.run(), functools.lru_cache()

Add function to get if we're x86_64_v2, so we can get the correct osg-release RPM URL.

Use 'yum swap' for upgrading the osg-release RPM.

+++ osgtest/library/core.py
@@ -1,6 +1,6 @@
 """Support and convenience functions for tests."""
 import contextlib
-import errno
+import functools
 import os
 import os.path
 import pwd
@@ -17,7 +17,7 @@ import traceback
 import socket
 import signal
 from shlex import quote as shell_quote
-from typing import Any, Dict, List, Union
+from typing import Any, Dict, List, Optional, Union
 
 from osgtest.library import osgunittest
 
@@ -56,7 +56,6 @@ options: Any = None
 _log = sys.stderr
 _log_filename = ""
 _last_log_had_output = True
-_el_release = ""
 
 SLURM_PACKAGES = ['slurm',
                   'slurm-slurmd',
@@ -126,7 +125,6 @@ def init_dummy():
     global options, _log
 
     options = DummyClass()
-    _log = DummyClass()
 
 
 def start_log():
@@ -192,16 +190,17 @@ def dump_log(outfile=None):
 
 def remove_log():
     """Removes the detailed log file; not for general use."""
-    os.remove(_log_filename)
+    if _log_filename:
+        os.remove(_log_filename)
+
 
 def get_stat(filename):
-    '''Return stat for 'filename', None if the file does not exist'''
+    """Return stat for 'filename', None if the file does not exist"""
     try:
         return os.stat(filename)
-    except OSError as exc:
-        if exc.errno == errno.ENOENT:
-            return None
-        raise
+    except FileNotFoundError:
+        return None
+
 
 def monitor_file(filename, old_stat, sentinel, timeout):
     """Monitors a file for the sentinel regex
@@ -368,8 +367,9 @@ def dependency_installed_rpms(a_dependency):
 def installed_rpms():
     """Returns the list of all installed packages."""
     command = ('rpm', '--query', '--all', '--queryformat', r'%{NAME}\n')
-    status, stdout, stderr = system(command, log_output=False, quiet=True)
-    return set(re.split(r'\s+', stdout.strip()))
+    proc = subprocess.run(command, stdout=subprocess.PIPE, encoding='latin-1')
+    return proc.stdout.splitlines()
+
 
 def rpm_regexp_is_installed(a_regexp):
     """Returns whether any RPM matches the provided regexp."""
@@ -639,26 +639,47 @@ def wait_for_file(filename, timeout):
 
     return False
 
-def el_release():
+
+@functools.lru_cache(1)
+def el_release() -> int:
     """Return the major version of the Enterprise Linux release the system is
     running. SL/RHEL/CentOS 6.x will return 6; SL/RHEL/CentOS 7.x will return
     7.
 
     """
-    global _el_release
-    if not _el_release:
+    try:
+        release_file = open("/etc/redhat-release", 'r')
         try:
-            release_file = open("/etc/redhat-release", 'r')
-            try:
-                release_text = release_file.read()
-            finally:
-                release_file.close()
-            match = re.search(r"release (\d)", release_text)
-            _el_release = int(match.group(1))
-        except (EnvironmentError, TypeError, ValueError) as e:
-            _log.write("Couldn't determine redhat release: " + str(e) + "\n")
-            sys.exit(1)
-    return _el_release
+            release_text = release_file.read()
+        finally:
+            release_file.close()
+        match = re.search(r"release (\d)", release_text)
+        return int(match.group(1))
+    except (EnvironmentError, TypeError, ValueError) as e:
+        _log.write("Couldn't determine redhat release: " + str(e) + "\n")
+        sys.exit(1)
+
+
+def osg_release_rpm() -> Optional[str]:
+    """
+    Return the name of the RPM providing osg-release, or None if no such
+    RPM is found.
+    """
+    proc = subprocess.run(
+        ['rpm', '-q', '--whatprovides', 'osg-release'],
+        stdout=subprocess.PIPE,
+        encoding='latin-1',
+    )
+    if proc.returncode != 0:
+        return None
+    return proc.stdout.strip()
+
+
+@functools.lru_cache(1)
+def is_x86_64_v2() -> bool:
+    """True if the OS (not necessarily the CPU) is Alma 10 x86_64_v2"""
+    proc = subprocess.run(['rpm', '-E', '%{x86_64_v2}'], stdout=subprocess.PIPE)
+    return proc.stdout.strip() == b'1'
 
 
 def osg_release(update_state=False):
@@ -667,10 +688,9 @@ def osg_release(update_state=False):
     """
     if not update_state and 'general.osg_release_ver' in state:
         return state['general.osg_release_ver']
-    try:
-        release = PackageVersion('osg-release')
-    except OSError:
-        release = PackageVersion('osg-release-itb')
+    release_rpm = osg_release_rpm()
+    assert release_rpm, "No osg-release RPM installed"
+    release = PackageVersion(release_rpm)
     state['general.osg_release_ver'] = release
     return release
